+++ packages/auth-utils/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stoxifyorg/auth-utils
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @stoxifyorg/logger@1.2.0
+  - @stoxifyorg/redis@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes