+++ packages/redis/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @stoxifyorg/redis
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @stoxifyorg/logger@1.2.0
+
 ## 1.0.1
 
 ### Patch Changes